Add test specific to minion start event
